Split out metric into multiple submodules
